Gave a User's interests a role in how compatible they are with another User

+++ src/main/java/UseCases/EditEdgeWeightUseCase.java
@@ -7,7 +7,7 @@ public class EditEdgeWeightUseCase {
 
     public void editUserEdgeWeight(User user1, User user2){
         CompatibilityFunction weightCalc = new CompatibilityFunction();
-        int weight = weightCalc.getCompatibility(user1, user2);
+        float weight = weightCalc.getCompatibility(user1, user2);
         UserGraph graph = new UserGraph();
         graph.getEdge(user1, user2).setWeight(weight);
     }
